requirements.txt
python-binance==1.0.19
pandas==2.0.3
numpy==1.24.3
TA-Lib==0.4.28
python-dotenv==1.0.0
colorlog==6.7.0
